Move log computation earlier
In linguistics, we tend to measure times in years, so heights are in the
order of magnitude of 100s to 1000s, not the 0.1My often found in
biology. In order to avoid overflow errors, the log should be taken
ASAP, and in fact this does not make the computation more expensive.